+++ examples/datasets/H5Ex_D_Transform.java
@@ -23,10 +23,10 @@ public class H5Ex_D_Transform {
     private static String RTRANSFORM = "x-1";
 
     private static void writeData() {
-        int file_id = -1;
-        int filespace_id = -1;
-        int dataset_id = -1;
-        int dxpl_id = -1;
+        long file_id = -1;
+        long filespace_id = -1;
+        long dataset_id = -1;
+        long dxpl_id = -1;
 
         long[] dims = { DIM_X, DIM_Y };
         int[][] dset_data = new int[DIM_X][DIM_Y];
@@ -130,9 +130,9 @@ public class H5Ex_D_Transform {
 
     private static void readData() {
 
-        int file_id = -1;
-        int dataset_id = -1;
-        int dxpl_id = -1;
+        long file_id = -1;
+        long dataset_id = -1;
+        long dxpl_id = -1;
         int[][] dset_data = new int[DIM_X][DIM_Y];
 
         //Open an existing file using the default properties.
